Arrange a data frame based on count, option, or total

This function arranges a data frame by count, option, or total, depending on the specified key. It uses dplyr::arrange() for arranging the data. The default arrangement is by descending count ("n").

Usage

enq_arrange(df, key = c("n", "option", "total", "default"))

arrange_total(df)

add_total(df)

Arguments

df

A dataframe to be arranged.

key

A character vector specifying the arrangement key. Possible values include:

"n": arrange by descending count ("n").

"option": arrange by the first column (assumed to be option).

"total": arrange by a total column (assumed to "total") in descending order (requires the data frame to already have a "total" column).

"default": arrange by descending count ("n") if the first column name does not contain "_ord", otherwise arrange by the first column.

Value

A dataframe.

Examples

enq_simple(dummy_data, "q1") |>
  print() |>
  enq_arrange(key = "option")
#> # A tibble: 21 × 2
#>    q1        n
#>    <chr> <int>
#>  1 b         6
#>  2 c         4
#>  3 m         4
#>  4 n         4
#>  5 u         4
#>  6 j         3
#>  7 r         3
#>  8 s         3
#>  9 a         2
#> 10 g         2
#> # ℹ 11 more rows
#> # A tibble: 21 × 2
#>    q1        n
#>    <chr> <int>
#>  1 a         2
#>  2 b         6
#>  3 c         4
#>  4 d         1
#>  5 g         2
#>  6 h         1
#>  7 j         3
#>  8 k         1
#>  9 l         2
#> 10 m         4
#> # ℹ 11 more rows
enq_simple(dummy_data, "q2_ord") |>
  print() |>
  enq_arrange(key = "n")
#> # A tibble: 5 × 2
#>   q2_ord        n
#>   <fct>     <int>
#> 1 always        4
#> 2 often        10
#> 3 sometimes    12
#> 4 rarely       14
#> 5 never        10
#> # A tibble: 5 × 2
#>   q2_ord        n
#>   <fct>     <int>
#> 1 rarely       14
#> 2 sometimes    12
#> 3 often        10
#> 4 never        10
#> 5 always        4
enq_cross(dummy_data, "q1", "q3_multi") |>
  print() |>
  enq_arrange(key = "total")
#> # A tibble: 21 × 8
#>    q1    q3_multi_ q3_multi_A q3_multi_B q3_multi_C q3_multi_D q3_multi_E total
#>    <chr>     <int>      <int>      <int>      <int>      <int>      <int> <int>
#>  1 b             0          6          4          3          4          4    21
#>  2 u             0          4          2          3          4          2    15
#>  3 c             1          3          3          2          2          3    14
#>  4 m             0          3          3          2          2          2    12
#>  5 g             0          2          2          1          1          2     8
#>  6 n             1          2          1          2          1          1     8
#>  7 l             0          1          1          1          2          1     6
#>  8 s             1          1          1          1          1          1     6
#>  9 t             0          1          1          1          2          1     6
#> 10 a             1          1          1          1          1          0     5
#> # ℹ 11 more rows
#> # A tibble: 21 × 8
#>    q1    q3_multi_ q3_multi_A q3_multi_B q3_multi_C q3_multi_D q3_multi_E total
#>    <chr>     <int>      <int>      <int>      <int>      <int>      <int> <int>
#>  1 b             0          6          4          3          4          4    21
#>  2 u             0          4          2          3          4          2    15
#>  3 c             1          3          3          2          2          3    14
#>  4 m             0          3          3          2          2          2    12
#>  5 g             0          2          2          1          1          2     8
#>  6 n             1          2          1          2          1          1     8
#>  7 l             0          1          1          1          2          1     6
#>  8 s             1          1          1          1          1          1     6
#>  9 t             0          1          1          1          2          1     6
#> 10 a             1          1          1          1          1          0     5
#> # ℹ 11 more rows