Create a cross tabulation of counts for two character columns

This function creates a cross tabulation of counts for two character columns in a data frame. It uses tidyr::separate_longer_delim() to split the columns if necessary, dplyr::count() to count the occurrences of each combination of unique values, and tidyr::pivot_wider() to reshape the data into a cross tabulation. The output is arranged by the values in the col_1 by default. An optional total can be added using the arrange_total() argument.

Usage

enq_cross(df, col_1, col_2, delim = ";", arrange_total = TRUE)

enq_cross_split(df, col_1, col_2, split = NULL, delim = ";")

Arguments

df

A dataframe containing the columns to be used for the cross tabulation.

col_1, col_2, split

A string for the name of the character column to be used in the cross tabulation.

delim

A string to be used for delimiter of values in the column (default is ";").

arrange_total

A logical value indicating whether to add a total. Defaults to TRUE.

Value

A data frame with the cross tabulation of counts. The first column contains the unique values from the col_1. The remaining columns are named based on the col_2 column values prefixed with "col_". Each cell contains the count of the corresponding combination of values from col_1 and col_2. An optional total are added if arrange_total is TRUE.

Examples

enq_cross(dummy_data, "q1", "q2_ord")
#> # A tibble: 21 × 7
#>    q1    q2_ord_always q2_ord_often q2_ord_sometimes q2_ord_rarely q2_ord_never
#>    <chr>         <int>        <int>            <int>         <int>        <int>
#>  1 b                 0            0                1             5            0
#>  2 c                 1            1                0             1            1
#>  3 m                 0            0                1             2            1
#>  4 n                 0            3                1             0            0
#>  5 u                 0            0                1             1            2
#>  6 j                 0            1                1             1            0
#>  7 r                 1            0                0             1            1
#>  8 s                 0            1                1             1            0
#>  9 a                 0            0                1             0            1
#> 10 g                 0            1                0             0            1
#> # ℹ 11 more rows
#> # ℹ 1 more variable: total <int>
enq_cross(dummy_data, "q1", "q3_multi")
#> # A tibble: 21 × 8
#>    q1    q3_multi_ q3_multi_A q3_multi_B q3_multi_C q3_multi_D q3_multi_E total
#>    <chr>     <int>      <int>      <int>      <int>      <int>      <int> <int>
#>  1 b             0          6          4          3          4          4    21
#>  2 u             0          4          2          3          4          2    15
#>  3 c             1          3          3          2          2          3    14
#>  4 m             0          3          3          2          2          2    12
#>  5 g             0          2          2          1          1          2     8
#>  6 n             1          2          1          2          1          1     8
#>  7 l             0          1          1          1          2          1     6
#>  8 s             1          1          1          1          1          1     6
#>  9 t             0          1          1          1          2          1     6
#> 10 a             1          1          1          1          1          0     5
#> # ℹ 11 more rows
enq_cross(dummy_data, "q2_ord", "q3_multi", arrange_total = FALSE)
#> # A tibble: 5 × 8
#>   q2_ord  q3_multi_ q3_multi_A q3_multi_B q3_multi_C q3_multi_D q3_multi_E total
#>   <fct>       <int>      <int>      <int>      <int>      <int>      <int> <int>
#> 1 always          1          3          1          2          1          2    10
#> 2 often           3          3          3          5          4          4    22
#> 3 someti…         2          9          3          6          5          4    29
#> 4 rarely          2         10         10          6          7          6    41
#> 5 never           0          7          7          4          7          5    30
enq_cross_split(dummy_data, "q1", "q2_ord", "q3_multi")
#> [[1]]
#> # A tibble: 8 × 6
#>   q1    q2_ord_always q2_ord_often q2_ord_sometimes q2_ord_rarely total
#>   <chr>         <int>        <int>            <int>         <int> <int>
#> 1 a                 0            0                1             0     1
#> 2 c                 1            0                0             0     1
#> 3 j                 0            0                0             1     1
#> 4 n                 0            1                0             0     1
#> 5 p                 0            1                0             0     1
#> 6 q                 0            1                0             0     1
#> 7 s                 0            0                1             0     1
#> 8 y                 0            0                0             1     1
#> 
#> $A
#> # A tibble: 18 × 7
#>    q1    q2_ord_always q2_ord_often q2_ord_sometimes q2_ord_rarely q2_ord_never
#>    <chr>         <int>        <int>            <int>         <int>        <int>
#>  1 b                 0            0                1             5            0
#>  2 u                 0            0                1             1            2
#>  3 c                 0            1                0             1            1
#>  4 m                 0            0                1             1            1
#>  5 g                 0            1                0             0            1
#>  6 n                 0            1                1             0            0
#>  7 a                 0            0                0             0            1
#>  8 d                 0            0                0             0            1
#>  9 h                 1            0                0             0            0
#> 10 k                 0            0                1             0            0
#> 11 l                 0            0                1             0            0
#> 12 r                 1            0                0             0            0
#> 13 s                 0            0                0             1            0
#> 14 t                 1            0                0             0            0
#> 15 v                 0            0                0             1            0
#> 16 x                 0            0                1             0            0
#> 17 y                 0            0                1             0            0
#> 18 z                 0            0                1             0            0
#> # ℹ 1 more variable: total <int>
#> 
#> $B
#> # A tibble: 15 × 7
#>    q1    q2_ord_always q2_ord_often q2_ord_sometimes q2_ord_rarely q2_ord_never
#>    <chr>         <int>        <int>            <int>         <int>        <int>
#>  1 b                 0            0                0             4            0
#>  2 c                 0            1                0             1            1
#>  3 m                 0            0                1             1            1
#>  4 g                 0            1                0             0            1
#>  5 u                 0            0                0             1            1
#>  6 a                 0            0                0             0            1
#>  7 d                 0            0                0             0            1
#>  8 l                 0            0                1             0            0
#>  9 n                 0            1                0             0            0
#> 10 p                 0            0                0             0            1
#> 11 r                 0            0                0             1            0
#> 12 s                 0            0                0             1            0
#> 13 t                 1            0                0             0            0
#> 14 v                 0            0                0             1            0
#> 15 z                 0            0                1             0            0
#> # ℹ 1 more variable: total <int>
#> 
#> $C
#> # A tibble: 16 × 7
#>    q1    q2_ord_always q2_ord_often q2_ord_sometimes q2_ord_rarely q2_ord_never
#>    <chr>         <int>        <int>            <int>         <int>        <int>
#>  1 b                 0            0                0             3            0
#>  2 u                 0            0                1             1            1
#>  3 c                 0            1                0             1            0
#>  4 m                 0            0                1             1            0
#>  5 n                 0            2                0             0            0
#>  6 a                 0            0                0             0            1
#>  7 d                 0            0                0             0            1
#>  8 g                 0            1                0             0            0
#>  9 h                 1            0                0             0            0
#> 10 j                 0            0                1             0            0
#> 11 k                 0            0                1             0            0
#> 12 l                 0            0                1             0            0
#> 13 r                 0            0                0             0            1
#> 14 s                 0            1                0             0            0
#> 15 t                 1            0                0             0            0
#> 16 z                 0            0                1             0            0
#> # ℹ 1 more variable: total <int>
#> 
#> $D
#> # A tibble: 14 × 7
#>    q1    q2_ord_always q2_ord_often q2_ord_sometimes q2_ord_rarely q2_ord_never
#>    <chr>         <int>        <int>            <int>         <int>        <int>
#>  1 b                 0            0                0             4            0
#>  2 u                 0            0                1             1            2
#>  3 c                 0            0                0             1            1
#>  4 l                 0            0                1             0            1
#>  5 m                 0            0                1             1            0
#>  6 t                 1            1                0             0            0
#>  7 a                 0            0                0             0            1
#>  8 d                 0            0                0             0            1
#>  9 g                 0            1                0             0            0
#> 10 n                 0            1                0             0            0
#> 11 p                 0            0                0             0            1
#> 12 s                 0            1                0             0            0
#> 13 x                 0            0                1             0            0
#> 14 z                 0            0                1             0            0
#> # ℹ 1 more variable: total <int>
#> 
#> $E
#> # A tibble: 13 × 7
#>    q1    q2_ord_always q2_ord_often q2_ord_sometimes q2_ord_rarely q2_ord_never
#>    <chr>         <int>        <int>            <int>         <int>        <int>
#>  1 b                 0            0                0             4            0
#>  2 c                 0            1                0             1            1
#>  3 g                 0            1                0             0            1
#>  4 m                 0            0                1             0            1
#>  5 u                 0            0                0             1            1
#>  6 d                 0            0                0             0            1
#>  7 h                 1            0                0             0            0
#>  8 j                 0            1                0             0            0
#>  9 l                 0            0                1             0            0
#> 10 n                 0            0                1             0            0
#> 11 s                 0            1                0             0            0
#> 12 t                 1            0                0             0            0
#> 13 y                 0            0                1             0            0
#> # ℹ 1 more variable: total <int>
#>