Count occurrences of values in a column

This function splits values in a character column by a delimiter and then counts the occurrences of each unique value. It uses tidyr::separate_longer_delim() to split the column and dplyr::count() to count the occurrences. The order of the output is determined by the presence of "_ord" in the column name. If "_ord" is present, the output is arranged by "option", otherwise it is arranged by "n" (number of occurrences).

Usage

enq_simple(df, col, delim = ";")

Arguments

df

A dataframe containing the column to be counted.

col

A string for column to be count.

delim

A string to be used for delimiter of values in the column (default is ";").

Value

A dataframe with two columns: "option" (the unique values after splitting) and "n" (the number of occurrences of each unique value).

Examples

enq_simple(dummy_data, "q1")
#> # A tibble: 21 × 2
#>    q1        n
#>    <chr> <int>
#>  1 b         6
#>  2 c         4
#>  3 m         4
#>  4 n         4
#>  5 u         4
#>  6 j         3
#>  7 r         3
#>  8 s         3
#>  9 a         2
#> 10 g         2
#> # ℹ 11 more rows
enq_simple(dummy_data, "q2_ord")
#> # A tibble: 5 × 2
#>   q2_ord        n
#>   <fct>     <int>
#> 1 always        4
#> 2 often        10
#> 3 sometimes    12
#> 4 rarely       14
#> 5 never        10
enq_simple(dummy_data, "q3_multi")
#> # A tibble: 6 × 2
#>   q3_multi     n
#>   <chr>    <int>
#> 1 "A"         32
#> 2 "B"         24
#> 3 "D"         24
#> 4 "C"         23
#> 5 "E"         21
#> 6 ""           8