Create summaries for multiple character columns in a data frame

This function creates summaries for multiple character columns in a data frame. It uses enq_simple to calculate the occurrences of split values for each column.

Usage

enq_summary(df, cols_except = "_comment$", delim = ";", simplify = FALSE)

enq_summary_bind(df)

Arguments

df

A dataframe containing the column to be counted.

cols_except

A regular expression pattern to exclude columns from the summary. Columns that do not match the pattern will be included. Defaults to "_comment$".

delim

A string to be used for delimiter of values in the column (default is ";").

simplify

A logical. If TRUE, bind elements of a list (default is FALSE).

Value

A list of dataframe. Names of the list elements are the same as the column names.

Examples

enq_summary(dummy_data)
#> $q1
#> # A tibble: 21 × 2
#>    q1        n
#>    <chr> <int>
#>  1 b         6
#>  2 c         4
#>  3 m         4
#>  4 n         4
#>  5 u         4
#>  6 j         3
#>  7 r         3
#>  8 s         3
#>  9 a         2
#> 10 g         2
#> # ℹ 11 more rows
#> 
#> $q2_ord
#> # A tibble: 5 × 2
#>   q2_ord        n
#>   <fct>     <int>
#> 1 always        4
#> 2 often        10
#> 3 sometimes    12
#> 4 rarely       14
#> 5 never        10
#> 
#> $q3_multi
#> # A tibble: 6 × 2
#>   q3_multi     n
#>   <chr>    <int>
#> 1 "A"         32
#> 2 "B"         24
#> 3 "D"         24
#> 4 "C"         23
#> 5 "E"         21
#> 6 ""           8
#> 
enq_summary(dummy_data, simplify = TRUE)
#> # A tibble: 32 × 3
#>    item  option     n
#>    <chr> <chr>  <int>
#>  1 q1    b          6
#>  2 q1    c          4
#>  3 q1    m          4
#>  4 q1    n          4
#>  5 q1    u          4
#>  6 q1    j          3
#>  7 q1    r          3
#>  8 q1    s          3
#>  9 q1    a          2
#> 10 q1    g          2
#> # ℹ 22 more rows